BAU: manually specify AWS_REGION on localstack configuration
The latest version of localstack is overriding the `AWS_REGION` property
passed in through the `env_file` of Docker compose. This causes the
resources to not line up with the other apps that share the same env
file.

It doesn't look like this has been resolved in the latest release so just
fix the `AWS_REGION` property for provisioning the resources until it starts
picking it up from the environment.