BAU: fix redis connection string for form builder adapter
The form builder adapter is starting up and healthy but fails to connect
to redis.

It looks like the redis client is failing to connect at startup and will
then fail to be used in any subsequent requests to the form builder runner.

Set the redis uri which looks like a param that was recently addeded to the
adapter. This allows the redis client to initialise as expected.