Hook up certificates bundle for API HTTPS
We are in the process of combining pre-award-stores and
pre-award-frontend.

The stores app runs over HTTP (insecure), but the frontend runs over
HTTPS (secure). Flask isn't able to serve both HTTP and HTTPS traffic
simultaneously, so we have decided to setup the API locally on HTTPS
(even though this isn't really representative of prod). We want HTTPS
locally so that we can legitimate authenticate with microsoft AD.

When the API is running on HTTPS with our self-signed cert, anything
that calls the API needs to have the root certificate available so that
it can trust/verify the connection. We do this by tweaking the local
certificate setup to store the root CA cert in `./certs`, and then
mounting this into the frontend container.

The frontend container then pulls in the root CA certificate to its
local store, and switches requests out from using the python `certifi`
bundle to the linux /etc/ssl/ca-certificates bundle, which now includes
our root CA cert.