BAU: Remove old DBs & local migration script
We no longer need the account or assessment store DBs as part of the
docker runner setup, as these have all been consolidated into pre-award.

Also removes the one-off local migration script which was used when we
migrated fund-store to pre-award-stores as a way for developers to
persist any funds they had set up locally. This script is now redundant.